Dodanie przetwarzania po id oferty x id z urlna widoku nidex

+++ PriceSafari/Models/AllegroPriceHistory.cs
@@ -46,6 +46,6 @@ namespace PriceSafari.Models
         public bool Promoted { get; set; }
         public bool Sponsored { get; set; }
 
-        public int IdAllegro { get; set; }
+        public long IdAllegro { get; set; }
     }
 }
